Removed GPR_ABSEIL_SYNC
